+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # TelegramBotAPI changelog
 
+## 18.2.3
+
+* `Core`:
+  * Add default middleware `ExceptionsThrottlerTelegramBotMiddleware`
+  * Make `TelegramBotMiddlewaresPipelinesHandler` to be default `TelegramBotPipelinesHandler`
+  * Make `DefaultKtorRequestsExecutor` now uses `runCatching` instead of `runCatchingSafely`
+
 ## 18.2.2
 
 * `Version`: